# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"opengame-api","description":"","license":{"name":""},"version":"0.1.0"},"paths":{"/auth/login":{"post":{"tags":[],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginSuccessResponse"}}}},"401":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginErrorResponse"}}}},"500":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginErrorResponse"}}}}}}},"/auth/register":{"post":{"tags":[],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterStatus"}}}},"409":{"description":"Username taken","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterStatus"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterStatus"}}}}}}},"/auth/status":{"get":{"tags":[],"operationId":"status","responses":{"200":{"description":"Get login status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoggedInStatus"}}}}}}},"/status/health":{"get":{"tags":[],"operationId":"health","responses":{"200":{"description":"Health check","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Health"}}}}}}}},"components":{"schemas":{"Health":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}}},"LoggedInStatus":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}}},"LoginErrorResponse":{"type":"object","required":["success","message"],"properties":{"message":{"type":"string"},"success":{"type":"boolean","example":false}}},"LoginRequest":{"type":"object","required":["username","password"],"properties":{"password":{"type":"string"},"username":{"type":"string"}}},"LoginSuccessResponse":{"type":"object","required":["success","bearer"],"properties":{"bearer":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"success":{"type":"boolean","example":true}}},"RegisterRequest":{"type":"object","required":["email","username","password"],"properties":{"email":{"type":"string"},"password":{"type":"string"},"username":{"type":"string"}}},"RegisterStatus":{"type":"object","required":["status"],"properties":{"status":{"type":"string"}}}}}}